refactor: 리팩토링 및 대시보드 출력 추가
- 기존 Task 조회 메서드 네이밍 개선 (`searchTask` → `findTaskById`, `searchTasks` → `findTasks`)
- Task 조회 로직 리팩토링 및 삭제 여부(isDeleted) 조건 추가
- `TaskReadRequest`에 managerId 필드 및 생성자 추가
- 대시보드 관련 기능 및 데이터 타입 변경 (`TaskOutline`)
- 대시보드 출력 로직 구현 (`findDashboard` in Repository/Service)
- 관련 Controller 및 Service 수정

+++ src/main/java/com/example/outsourcingproject/domain/dashboard/dto/TaskOutline.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class TaskOutline {
-    private int totalCount;
-    private int todoCount;
-    private int inProgressCount;
-    private int doneCount;
+    private Long totalCount;
+    private Long todoCount;
+    private Long inProgressCount;
+    private Long doneCount;
     private double completionRate;
-    private int overdueCount;
+    private Long overdueCount;
 }
